requirements.txt
flask
flask_cors
tensorflow
Pillow
numpy
streamlit
requests